+++ tests/test_exploitation.py
@@ -3,9 +3,10 @@ import os
 import pyexcel as pe
 
 from .nose_tools import eq_
+import unittest
 
 
-class TestIrregularities:
+class TestIrregularities(unittest.TestCase):
     def setUp(self):
         self.testfile = "test.xlsm"
         self.content = [
@@ -33,7 +34,7 @@ class TestIrregularities:
             os.unlink(self.testfile)
 
 
-class TestIrregularities2:
+class TestIrregularities2(unittest.TestCase):
     def setUp(self):
         self.testfile = "test.xlsm"
         self.content = {
@@ -84,7 +85,7 @@ class TestIrregularities2:
             os.unlink(self.testfile)
 
 
-class TestIrregularities10:
+class TestIrregularities10(unittest.TestCase):
     def setUp(self):
         self.testfile = "test.xlsm"
         self.content = [
@@ -121,7 +122,7 @@ class TestIrregularities10:
             os.unlink(self.testfile)
 
 
-class TestIrregularities3:
+class TestIrregularities3(unittest.TestCase):
     def setUp(self):
         self.testfile = "test.csv"
         self.content = []
@@ -141,7 +142,7 @@ class TestIrregularities3:
             os.unlink(self.testfile)
 
 
-class TestMultiSheetReader:
+class TestMultiSheetReader(unittest.TestCase):
     def setUp(self):
         self.testfile = "file_with_an_empty_sheet.xlsx"
 
@@ -152,7 +153,7 @@ class TestMultiSheetReader:
         assert r.number_of_sheets() == 3
 
 
-class TestEmptyCSVFile:
+class TestEmptyCSVFile(unittest.TestCase):
     def setUp(self):
         self.testfile = os.path.join("tests", "fixtures", "empty.csv")
 
@@ -161,7 +162,7 @@ class TestEmptyCSVFile:
         assert r.number_of_columns() == 0
 
 
-class TestNonUniformCSVFile:
+class TestNonUniformCSVFile(unittest.TestCase):
     def setUp(self):
         self.testfile = os.path.join(
             "tests",
@@ -184,7 +185,7 @@ class TestNonUniformCSVFile:
         eq_(result, r.array)
 
 
-class TestMergeFiles:
+class TestMergeFiles(unittest.TestCase):
     def test_merge_two_csv_files(self):
         file1 = os.path.join(
             "tests",
